% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_wc.R
\name{ef_wc}
\alias{ef_wc}
\title{ef_wc}
\usage{
ef_wc(x, dc_threshold, c_weeks, ef)
}
\arguments{
\item{x}{where denotes the numerical vector of the number of cases per unit time.}

\item{dc_threshold}{where is the threshold for the number of cases.}

\item{c_weeks}{is the consecutive week.}

\item{ef}{is the epifeature. The options are ds3, ds6, dsmed and ppw.}
}
\value{
This function returns a vector numeric.
}
\description{
Epidemiological characteristics with cases
}
\details{
The function and arguments were based on the work of Almeida et al. (2022) for a more detailed review. Almeida IF, Lana RM, Codeco CT. 2022. How heterogeneous is the dengue transmission profile in Brazil? A study in six Brazilian states. PLO Neglected Tropical Diseases.
}
\examples{
ef_wc(x = c(8, 15, 20, 0, 0, 0, 0, 5, 9 ,12), ef = "freq", dc_threshold = 5, c_weeks = 3)
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
