% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_woc.R
\name{ef_woc}
\alias{ef_woc}
\title{ef_woc}
\usage{
ef_woc(x, dc_threshold, c_weeks, ef)
}
\arguments{
\item{x}{where denotes the numerical vector of the number of cases per unit time.}

\item{dc_threshold}{where is the threshold for the number of cases.}

\item{c_weeks}{is the consecutive week.}

\item{ef}{is the epifeature. The options are ds3, ds6,dsmax, & dsmed.}
}
\value{
This function returns a vector numeric.
}
\description{
Epidemiological characteristics without cases
}
\details{
The function and arguments were based on the work of Almeida et al. (2022) for a more detailed review. Almeida IF, Lana RM, Codeco CT. 2022. How heterogeneous is the dengue transmission profile in Brazil? A study in six Brazilian states. PLO Neglected Tropical Diseases.
}
\examples{
ef_woc(x = c(2, 0, 0, 0, 3, 5, 0, 0, 0, 0,2), dc_threshold = 0, c_weeks = 3, ef = "ds3")
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
