% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epifeatures-package.R
\docType{package}
\name{epifeatures-package}
\alias{epifeatures}
\alias{epifeatures-package}
\title{epifeatures: Epidemiological characteristics of the dengue time series}
\description{
extracts epi-features from dengue datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fdzul/epifeatures}
  \item \url{https://fdzul.github.io/features}
  \item Report bugs at \url{https://github.com/fdzul/features/issues}
}

}
\author{
\strong{Maintainer}: Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com} (\href{https://orcid.org/0000-0002-2085-9141}{ORCID})

Other contributors:
\itemize{
  \item Fabian Correa-Morales \email{fabiancorrea@msn.com} (\href{https://orcid.org/0000-0002-6193-1242}{ORCID}) [contributor]
}

}
\keyword{internal}
